+++ lua/yabs/conditions/init.lua
@@ -2,7 +2,7 @@ local M = {}
 
 function M.file_exists(file)
   return function()
-    return require('yabs.util').file_exists(file)
+    return require('yabs.utils').file_exists(file)
   end
 end
