+++ lua/yabs/language.lua
@@ -3,7 +3,7 @@ local Language = {}
 local Task = require('yabs.task')
 
 local output_types = require('yabs.outputs')
-local util = require('yabs.util')
+local utils = require('yabs.utils')
 
 function Language:new(args)
   local state = {
@@ -33,13 +33,13 @@ function Language:setup(parent, args)
   if args then
     if args.default == true then
       parent.default_language = self
-      util.notify(
+      utils.notify(
         'yabs: deprecation notice: `default` and `override` languages are superceded by global tasks',
         vim.log.levels.WARN
       )
     end
     if args.override == true then
-      util.notify(
+      utils.notify(
         'yabs: deprecation notice: `default` and `override` languages are superceded by global tasks',
         vim.log.levels.WARN
       )
@@ -78,7 +78,7 @@ function Language:has_task(task)
   end
   -- TODO: remove this, tasks as tables is no longer supported
   if type(task) == 'table' then
-    util.notify('yabs: tasks as tables are no longer supported', vim.log.levels.ERROR)
+    utils.notify('yabs: tasks as tables are no longer supported', vim.log.levels.ERROR)
   end
   return false
 end
@@ -89,7 +89,7 @@ function Language:run_task(task, opts)
     self.tasks[task]:run(opts)
   elseif type(task) == 'table' then
     -- TODO: remove this, tasks as tables is no longer supported
-    util.notify('yabs: tasks as tables are no longer supported', vim.log.levels.ERROR)
+    utils.notify('yabs: tasks as tables are no longer supported', vim.log.levels.ERROR)
   end
 end
