+++ lua/yabs/outputs/none.lua
@@ -1,5 +1,5 @@
 local function none(cmd)
-  require('yabs.util').async_command(cmd)
+  require('yabs.utils').async_command(cmd)
 end
 
 local Output = require('yabs.output')
